requirements.txt
pandas==2.1.4
numpy==1.26.2
matplotlib==3.8.2
seaborn==0.13.0
scipy==1.11.4
jupyter==1.0.0
ipykernel==6.27.1
python-pptx==0.6.23
